Pass a JS object instead of the C++->QML one to the infoMovie
The C++->QML object as a part of the model that is only used in the grid view
/ list view component so it's destroyed right after the stack use the .replace
method (when it's doing so, it's destroying the old component and then any
reference that uses the model, so item is also destroyed). When the object
passed was the item of the model, it was conisdered reupdated (deleted in fact)
right after the infoMovie componenent is displayed and thus was sneding warning
as it was trying to get some properties (title and cover) from a null element.
So to solve this, the values taht are interesting are now saved into a JS object
which is not destroyed when the component is replaced in the stack.