List/Grid view is now a child of the category.
Before, the loader that choose between Albums/Artists/... was a child of
the loader that choose between List and Grid view. By inverting both, it allows
to customise each elements of the category independantly.
The precedent version was stuck at either a GridView or an ExpandGridView for all
category, not a mix of thoses.